Lógico inicial começando

+++ pid.py
@@ -1,14 +1,13 @@
 from ctypes import * 
 import os
 
-so_file = os.path.join(os.path.dirname(__file__),'src/pid/pid.so') 
+so_file = os.path.join(os.path.dirname(__file__),'src/pid/pid.so')
 
 pid_functions = CDLL(so_file)
 pid_functions.pid_configura_constantes.argtypes = [c_double,c_double,c_double]
-pid_functions.pid_atualiza_referencia.argtypes = [c_float]
+pid_functions.pid_atualiza_referencia.argtypes = [c_double]
 pid_functions.pid_controle.argtypes = [c_double]
 pid_functions.pid_controle.restype = c_double
-
 
 def pid_configura_constantes(kp,ki,kd):
     pid_functions.pid_configura_constantes(kp,ki,kd)
